Parallel test

+++ src/test/java/com/legion/tests/TestBase.java
@@ -187,33 +187,36 @@ public abstract class TestBase {
 
     @BeforeMethod(alwaysRun = true)
     protected void initTestFramework(Method method, ITestContext context) throws AWTException, IOException, APIException, JSONException {
-        Date date=new Date();
-        SimpleDateFormat formatter = new SimpleDateFormat("dd MMMM yyyy");
-        String testName = ExtentTestManager.getTestName(method);
-        String ownerName = ExtentTestManager.getOwnerName(method);
-        String automatedName = ExtentTestManager.getAutomatedName(method);
-        enterpriseName =  SimpleUtils.getEnterprise(method);
-        String platformName =  ExtentTestManager.getMobilePlatformName(method);
-        List<String> categories =  new ArrayList<String>();
-        categories.add(getClass().getSimpleName());
-        List<String> enterprises =  new ArrayList<String>();
-        enterprises.add(enterpriseName);
-        ExtentTestManager.createTest(getClass().getSimpleName() + " - "
-                + " " + method.getName() + " : " + testName + ""
-                + " [" + ownerName + "/" + automatedName + "/" + platformName + "]", "", categories);
-        extent.setSystemInfo(method.getName(), enterpriseName.toString());
-        if (testRailRunId==null){
-            testRailRunId = 0;
+        try {
+            Date date = new Date();
+            String testName = ExtentTestManager.getTestName(method);
+            String ownerName = ExtentTestManager.getOwnerName(method);
+            String automatedName = ExtentTestManager.getAutomatedName(method);
+            enterpriseName = SimpleUtils.getEnterprise(method);
+            String platformName = ExtentTestManager.getMobilePlatformName(method);
+            List<String> categories = new ArrayList<String>();
+            categories.add(getClass().getSimpleName());
+            List<String> enterprises = new ArrayList<String>();
+            enterprises.add(enterpriseName);
+            ExtentTestManager.createTest(getClass().getSimpleName() + " - "
+                    + " " + method.getName() + " : " + testName + ""
+                    + " [" + ownerName + "/" + automatedName + "/" + platformName + "]", "", categories);
+            extent.setSystemInfo(method.getName(), enterpriseName.toString());
+            if (testRailRunId == null) {
+                testRailRunId = 0;
+            }
+            if (testRailReportingFlag != null) {
+                TestRailOperation.addNUpdateTestCaseIntoTestRail(testName, context);
+                MyThreadLocal.setTestResultFlag(false);
+                MyThreadLocal.setTestSkippedFlag(false);
+            }
+            setCurrentMethod(method);
+            setBrowserNeeded(true);
+            setCurrentTestMethodName(method.getName());
+            setSessionTimestamp(date.toString().replace(":", "_").replace(" ", "_"));
+        } catch (Exception e) {
+            SimpleUtils.fail("Error encountered when initing framework: " + e.getMessage(), false);
         }
-        if(testRailReportingFlag!=null){
-            TestRailOperation.addNUpdateTestCaseIntoTestRail(testName,context);
-            MyThreadLocal.setTestResultFlag(false);
-            MyThreadLocal.setTestSkippedFlag(false);
-        }
-        setCurrentMethod(method);
-        setBrowserNeeded(true);
-        setCurrentTestMethodName(method.getName());
-        setSessionTimestamp(date.toString().replace(":", "_").replace(" ", "_"));
     }
 
 
@@ -340,8 +343,10 @@ public abstract class TestBase {
         TestRailOperation.addResultForTest();
         if (Boolean.parseBoolean(propertyMap.get("close_browser"))) {
             try {
-                getDriver().manage().deleteAllCookies();
-                getDriver().quit();
+                if (getDriver() != null) {
+                    getDriver().manage().deleteAllCookies();
+                    getDriver().quit();
+                }
             } catch (Exception exp) {
                 Reporter.log("Error closing browser");
             }
@@ -555,10 +560,10 @@ public abstract class TestBase {
         }
         newShiftPage.clickOnCreateOrNextBtn();
         if (assignment.equals(ScheduleTestKendraScott2.staffingOption.AssignTeamMemberShift.getValue())) {
-            shiftOperatePage.switchSearchTMAndRecommendedTMsTab();
             if (tmName != null && !tmName.isEmpty()) {
                 newShiftPage.searchTeamMemberByName(tmName);
             } else {
+                shiftOperatePage.switchSearchTMAndRecommendedTMsTab();
                 for (int i = 0; i < shiftPerDay; i++) {
                     selectedTMs.add(newShiftPage.selectTeamMembers());
                 }
